LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/navigate-history-push-same-url-cross-document-expected.txt



FAIL navigate() to the current URL with history: 'push' and allow it to go cross document assert_equals: expected (string) "push" but got (undefined) undefined
